engine_rocks: move tests to engine.rs
